refactor: Split opinion views into multiple files
