+++ opinions/__init__.py
@@ -22,7 +22,7 @@
 #
 from .constants import (
     OPINION_ID_ROUTE_NAME, OPINION_NEW_ROUTE_NAME, OPINION_SLUG_ROUTE_NAME,
-    OPINION_PREVIEW_ID_ROUTE_NAME
+    OPINION_PREVIEW_ID_ROUTE_NAME, OPINION_STATUS_ID_ROUTE_NAME
 )
 
 __all__ = [
@@ -30,4 +30,5 @@ __all__ = [
     'OPINION_NEW_ROUTE_NAME',
     'OPINION_SLUG_ROUTE_NAME',
     'OPINION_PREVIEW_ID_ROUTE_NAME',
+    'OPINION_STATUS_ID_ROUTE_NAME',
 ]
